properly calculate GlyphInfos in textshaper
